==> 1_data_obtain_wrangling.ipynb <==
# Obtain Data & Wrangling

Data can be downloaded from Kaggle website (https://www.kaggle.com/c/instacart-market-basket-analysis/data).

Once it is downloaded, unzip and place all the extracted files to  "/data" folder (created one if it doesn't exist) of the current main folder.

Data can also be obtained directly from Instacart website (https://www.instacart.com/datasets/grocery-shopping-2017)

### [Data Dictionary](https://gist.github.com/jeremystan/c3b39d947d9b88b3ccff3147dbcf6c6b)

*“The Instacart Online Grocery Shopping Dataset 2017”, Accessed from https://www.instacart.com/datasets/grocery-shopping-2017 on <date>*

import pandas as pd
import numpy as np
from functools import reduce
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

### Examine Data Structure and check missing values

aisles = pd.read_csv('./data/aisles.csv')
dept = pd.read_csv('./data/departments.csv')
orders = pd.read_csv('./data/orders.csv')
products = pd.read_csv('./data/products.csv')
orders_p = pd.read_csv('./data/order_products__prior.csv')
orders_tr = pd.read_csv('./data/order_products__train.csv')

def examine(dataset):
    display(dataset.head())
    print(dataset.shape)
    print(dataset.info())
    display('missing data check', dataset.isnull().sum())

examine(aisles)
examine(dept)
examine(products)

examine(orders)

examine(orders_p)

examine(orders_tr)

Most of the attributes had been converted to int64. No major missing values except 'days_since_prior_order', which is expected when order number is one

orders.loc[orders['days_since_prior_order'].isnull()==True]['order_number'].value_counts()

**Conclusion**: data are already clean and they are in the right data type. No further cleaning is needed.
