==> 3_ML_features.ipynb <==
# Machine Learning - Feature Creations

We will use all the prior order information to generate features, and use the train data to create the target variables.

From exploratory analysis, we learned that purchase pattern at the department level is pretty consistence, and the reorder pattern are similar to orders created by users before. Also, reorder on average fall under 15 days for new clients, and gradually decrease over time to 2-4 days. Hence, we will test the following features
- product purchase frequency
- product purchase recency

### Import modules and data

import pandas as pd
import numpy as np
import gc
from pycaret.classification import *

# aisles = pd.read_csv('./data/aisles.csv')
# dept = pd.read_csv('./data/departments.csv')
orders = pd.read_csv('./data/orders.csv')
products = pd.read_csv('./data/products.csv')
orders_p = pd.read_csv('./data/order_products__prior.csv')
orders_tr = pd.read_csv('./data/order_products__train.csv')

prior_order = orders.query('eval_set == "prior"')
train_order = orders.query('eval_set == "train"')
test_order = orders.query('eval_set == "test"')

products.head()

print(orders.shape)
display(orders.head())
print(orders_p.shape)
print(orders_tr.shape)
display(orders_p.head(), orders_tr.head())

### Feature Creations

- last purchase: whether a particular product was purchased in the latest order by user
- last 3 purchase: number of times a product appear in the last 3 purchases
- the average duration (in terms of days) that item under a certain department was purchased
- how many times an item has been purchased by the user
- how often that item appear in user purchase history, in form of percentage

##### Obtain a list of last purchase order by user

last_purchase = prior_order[prior_order['order_number'] == prior_order.groupby(
    ['user_id'])['order_number'].transform('max')]
last_purchase_list = last_purchase['order_id'].tolist()

##### Generating last purchase feature

order_detail = orders_p.merge(prior_order[['order_id', 'user_id',
                            'order_number', 'days_since_prior_order']],
                              on=['order_id'], how='left')
order_detail.loc[order_detail['order_id'].isin(last_purchase_list),
                 'last_purchase'] = 1
order_detail.loc[~order_detail['order_id'].isin(last_purchase_list),
                 'last_purchase'] = 0

##### Generating average days of department/item reordering

The feature is to cover purchase pattern of a user at the department level. A user may switch between different brands of product (e.g. snacks) among his/her purchase history.

Steps on getting the average days of product reordering

- Get a cumulative sum of days of each order, create column (cumsum)
- apply the cumsum at the order level to products within each order
- get department id from the product id, and map it back to cumsum of each order
- use groupby then .diff() at the user/department level to get the absolute date lag between each order containing items within certain departments

ord_duration = orders[['user_id', 'order_number', 'days_since_prior_order']]
ord_duration['cumsum'] = ord_duration.groupby(['user_id'])['days_since_prior_order'].cumsum()
ord_duration = ord_duration.drop('days_since_prior_order', axis=1)
ord_duration = ord_duration.dropna()

gc.collect()
get_dept = order_detail[['product_id', 'user_id', 'order_number']]
get_dept['dummy'] = 1# dummy column to allow groupby in next step
get_dept = pd.merge(get_dept, products[['product_id', 'department_id']],
                     on='product_id', how='left')
get_dept = get_dept.groupby(['user_id', 'department_id',
                               'order_number']).agg({'dummy': 'count'})
get_dept = get_dept.reset_index()
get_dept = pd.merge(get_dept, ord_duration,
                    on=['user_id', 'order_number'], how='left')

get_dept.head()

dept_lag = get_dept.copy()
dept_lag['lag'] = dept_lag.groupby(['user_id',
                                      'department_id'])['cumsum'].diff()

final_lag = dept_lag[['user_id', 'department_id',
                       'order_number', 'lag']]

Add lag features back to the order_detail dataframe.

order_detail = pd.merge(order_detail,
                        products[['product_id', 'department_id']],
                        on='product_id', how='left')
order_detail = pd.merge(order_detail, final_lag,
                        on=['user_id', 'department_id', 'order_number'], how='left')

order_detail.head()

##### Generate item purchased count and percentage of item appeared in user purchase history

user_df = order_detail.groupby(['user_id', 'product_id']).agg({
    'order_number': 'max', 'last_purchase': 'max', 'reordered': 'sum', 'lag': 'mean'}).fillna(0)

user_df.head()

user_df['product_appear'] = (user_df['reordered']+1)/user_df['order_number']

user_order_cnt = orders.groupby(['user_id', 'order_number']).agg({
    'days_since_prior_order': 'sum'})

user_df.head()

##### Generate consecutive purchase features for the last 3 orders, by user and by item

# Get the last 3 purchases
user_order_cnt_1 = user_order_cnt.groupby(level=0).apply(lambda df: df[-4:])
user_order_cnt_1.index = user_order_cnt_1.index.droplevel(0)

last_three = user_order_cnt_1.reset_index().drop(
    'days_since_prior_order', axis=1)

product_last_three = pd.merge(last_three, order_detail,
                              on=['user_id', 'order_number'], how='inner')
product_cnt = product_last_three.groupby(['user_id', 'product_id']).agg(
    {'product_id': 'count'})

##### Forming final training dataframe

temp_df = pd.merge(user_df, product_cnt, left_index=True,
                   right_index=True, how='left')

temp_df = temp_df.rename(columns={'product_id': 'buy_cnt'})

temp_df.reset_index(inplace=True)

# save copies for ease of access on other model training scearios
temp_df.to_csv('./data/df.csv', index=True)

# getting user id within the train dataset section
train_id = train_order.user_id.unique().tolist()

# filter out the training set user id from the temp dataframe
temp_train_df = temp_df.loc[temp_df['user_id'].isin(train_id)]

Generating target features from the train data

train_target = pd.merge(train_order[['order_id', 'user_id']], orders_tr,
                        on=['order_id'], how='left')
train_train = train_target.drop(['order_id', 'add_to_cart_order'], axis=1)
train_target = train_target.rename(columns={'reordered': 'target'})

# set to 1 as these are the products brought in the next purchase
train_target['target'] = 1

Add target feature to the training dataframe, and drop order_number column

train_df = pd.merge(temp_train_df, train_target, on=['user_id', 'product_id'],
                    how='outer').fillna(0)
train_df = train_df.set_index(['user_id', 'product_id'])

# columns that won't be needed for final features
train_df.drop(['order_number', 'order_id', 'add_to_cart_order'],
              axis=1, inplace=True)

Set target to integer to ensure model recognize that as binary variable

train_df = train_df.astype({'target': 'int'})

train_df.head()

train_df.to_csv('./data/df.csv', index=True)

##### Create separate formatted test dataset for ease of access later.

test_id = test_order.user_id.unique().tolist()
df_for_submit = temp_df.loc[temp_df['user_id'].isin(test_id)].set_index(
    ['user_id', 'product_id']).fillna(0)

df_for_submit.to_csv('./data/df_submit.csv')

[Modeling Part 1](https://github.com/sittingman/instacart_product_repurchase/blob/master/4_ML_model_p1.ipynb)
<br>
<br>
[Modeling Part 2](https://github.com/sittingman/instacart_product_repurchase/blob/master/4_ML_model_p2.ipynb)
- repeat modeling part 1 but dropping 'lag' feature and re-access best model fitting
