==> 4_ML_model_p2.ipynb <==
# Machine Learning - Modeling (Part 2)

This section is to repeat the machine learning process by excluding "lags" features.

import pandas as pd
from pycaret.classification import *

orders = pd.read_csv('./data/orders.csv')
prior_order = orders.query('eval_set == "prior"')
train_order = orders.query('eval_set == "train"')
test_order = orders.query('eval_set == "test"')

df = pd.read_csv('./data/df.csv')

df.head()

df = df.set_index(['user_id', 'product_id'])

#### Setup

exp = setup(df, target='target', categorical_features=['last_purchase'],
            ignore_features=['lag'], train_size=.8)

#### Compare models

See if any models standout from the F1 score perspective

compare_models(blacklist=['knn', 'ridge', 'svm', 'lda', 'nb', 'qda', 'et', 'catboost'],
               fold=2, round=3, sort='F1')

**Observation**: Taking out the feature 'lag' overall F1 score stay relatively the same. Extreme Gradient Boosting (Xgboost) become the best performance mode. Trees-based models such as Random Forecast see higher performance boost. We will test out Xgboost and Random Forecast to see if we can achieve the same level of without the 'lag" feature.

### Save all the models for final test data submission

path = './data/save_model/'

def final_save(model, name):
    """ create the final model and save that as pkl file

    Args:
    model: model created with the sample data

    Returns:
    finalize model using all the data
    """

    final = finalize_model(model)
    save_model(final, model_name=path+name)

### Tune Model

We will go straight to the tune model part to optimize for best F1 score.

rfc_tune = tune_model('rf', fold=2, optimize='F1')
final_save(rfc_tune, 'rfc_tune')

xbc_tune = tune_model('xgboost', fold=2, n_iter=5, optimize='F1')

final_save(xbc_tune, 'xbc_tune')

#### Kaggle Score (tuned model)

| Model | F1 |
|------| ----|
| Xgboost | 0.36049 |
| Random Forecast| 0.36025|

Both models achieved higher F1 score when 'lag' feature is dropped.

### Data Submission

Go to [data_sumbit](https://github.com/sittingman/instacart_product_repurchase/blob/master/5_data_submit.ipynb) workbook.
